+++ src/teritorio/__init__.py
@@ -1,4 +1,4 @@
 from teritorio.main import Countries, Currencies
 
 __all__ = ["Countries", "Currencies"]
-__version__ = "2021.10.22"
+__version__ = "2023.06.13"